Heightmaps

+++ src/main/java/org/collabthings/math/CTMath.java
@@ -7,22 +7,22 @@ import waazdoh.common.WObject;
 public class CTMath {
 
 	public static double radToDegrees(double angle) {
-		angle = limitAngle(angle);
+		angle = limitAngle((float) angle);
 
 		return 360.0 * angle / (2 * Math.PI);
 	}
 
-	public static double limitAngle(double angle) {
+	public static float limitAngle(double angle) {
 		while (angle < -2 * Math.PI)
 			angle += 2 * Math.PI;
 		while (angle > 2 * Math.PI)
 			angle -= 2 * Math.PI;
-		return angle;
+		return (float) angle;
 	}
 
 	public static double degreesToRad(double na) {
 		double r = 2 * Math.PI * na / 360.0;
-		return limitAngle(r);
+		return limitAngle((float) r);
 	}
 
 	public static Vector3f parseVector(WObject o) {
